Moved detector filters in new sub-package
