+++ src/proc_image_processing/filters/detectors/center_coffin_detector.h
