+++ src/proc_image_processing/filters/detectors/fence_detector.h
