+++ src/proc_image_processing/filters/detectors/gate_detector.h
