+++ src/proc_image_processing/filters/detectors/handle_detector.h
