+++ src/proc_image_processing/filters/detectors/pipe_angle_detector.h
