+++ src/proc_image_processing/filters/detectors/vampire_body_detector.h
