+++ src/proc_image_processing/filters/detectors/vampire_torpedoes_close_detector.h
