+++ src/proc_image_processing/server/filter_factory.h
@@ -11,17 +11,17 @@
 #include <proc_image_processing/filters/bilateral_filter.h>
 #include <proc_image_processing/filters/blurr.h>
 #include <proc_image_processing/filters/canny.h>
-#include <proc_image_processing/filters/center_coffin_detector.h>
+#include <proc_image_processing/filters/detectors/center_coffin_detector.h>
 #include <proc_image_processing/filters/contrast_brightness.h>
 #include <proc_image_processing/filters/convex_hull.h>
 #include <proc_image_processing/filters/deep_2019.h>
 #include <proc_image_processing/filters/dilate.h>
 #include <proc_image_processing/filters/equalize.h>
 #include <proc_image_processing/filters/erode.h>
-#include <proc_image_processing/filters/fence_detector.h>
+#include <proc_image_processing/filters/detectors/fence_detector.h>
 #include <proc_image_processing/filters/filter.h>
-#include <proc_image_processing/filters/gate_detector.h>
-#include <proc_image_processing/filters/handle_detector.h>
+#include <proc_image_processing/filters/detectors/gate_detector.h>
+#include <proc_image_processing/filters/detectors/handle_detector.h>
 #include <proc_image_processing/filters/hough_line.h>
 #include <proc_image_processing/filters/hsv_threshold.h>
 #include <proc_image_processing/filters/image_accumulator.h>
@@ -31,7 +31,7 @@
 #include <proc_image_processing/filters/mission_test_fake_string.h>
 #include <proc_image_processing/filters/morphology.h>
 #include <proc_image_processing/filters/original_image.h>
-#include <proc_image_processing/filters/pipe_angle_detector.h>
+#include <proc_image_processing/filters/detectors/pipe_angle_detector.h>
 #include <proc_image_processing/filters/rotate.h>
 #include <proc_image_processing/filters/remove_mask.h>
 #include <proc_image_processing/filters/schar_adding.h>
@@ -44,9 +44,9 @@
 #include <proc_image_processing/filters/test_filter.h>
 #include <proc_image_processing/filters/threshold.h>
 #include <proc_image_processing/filters/threshold_between.h>
-#include <proc_image_processing/filters/vampire_body_detector.h>
-#include <proc_image_processing/filters/vampire_torpedoes_detector.h>
-#include <proc_image_processing/filters/vampire_torpedoes_close_detector.h>
+#include <proc_image_processing/filters/detectors/vampire_body_detector.h>
+#include <proc_image_processing/filters/detectors/vampire_torpedoes_detector.h>
+#include <proc_image_processing/filters/detectors/vampire_torpedoes_close_detector.h>
 #include <proc_image_processing/filters/white_filter.h>
 #include <proc_image_processing/filters/white_noise_takedown.h>
 #include <memory>
